requirements.txt
python-dotenv==0.15.0
requests==2.25.0
urllib3==1.26.2
